services/backend/requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
sqlalchemy==2.0.29
python-multipart
python-jose
passlib[bcrypt]
websockets
asyncio==0.1.1